The dissasembler part is mostly done.  It fills out the fucntion calls and we are decoding arrays and objects.   I even have it pulling add operations and doing things like "1 == global.flag[3]".
This might be it for the dissasembler.  Its mostly done and eveything after this is about clean up.  I only wanted the dissasembler good enough for somone to eyeball how things are working.  Anything more will be pointless.

AFter this going to work on the texture/font converter a bit better.  Clean it up for sure.